+++ packages/localnet/src/index.ts
@@ -17,9 +17,9 @@ import { handleOnZEVMWithdrawn } from "./handleOnZEVMWithdrawn";
 import { createToken } from "./createToken";
 import { handleOnZEVMWithdrawnAndCalled } from "./handleOnZEVMWithdrawnAndCalled";
 import { handleOnEVMDepositedAndCalled } from "./handleOnEVMDepositedAndCalled";
-import { setupSolana } from "./setupSolana";
+import { solanaSetup } from "./solanaSetup";
 import * as anchor from "@coral-xyz/anchor";
-import Gateway_IDL from "./solana/idl/gateway.json";
+import { solanaMonitorTransactions } from "./solanaMonitorTransactions";
 
 const FUNGIBLE_MODULE_ADDRESS = "0x735b14BB79463307AAcBED86DAf3322B1e6226aB";
 
@@ -249,84 +249,6 @@ const deployProtocolContracts = async (
     uniswapRouterInstance,
   };
 };
-async function monitorOnlyNewTransactions(program: any, connection: any) {
-  console.log(
-    `Monitoring new transactions for program: ${program.programId.toBase58()}`
-  );
-
-  let lastSignature: string | undefined = undefined;
-
-  setInterval(async () => {
-    let signatures;
-    try {
-      signatures = await connection.getSignaturesForAddress(
-        program.programId,
-        { limit: 10 },
-        "confirmed"
-      );
-
-      if (signatures.length === 0) return;
-
-      const newSignatures = [];
-
-      for (const signatureInfo of signatures) {
-        if (signatureInfo.signature === lastSignature) {
-          break;
-        } else {
-          newSignatures.push(signatureInfo);
-        }
-      }
-
-      if (newSignatures.length === 0) return;
-
-      for (const signatureInfo of newSignatures.reverse()) {
-        try {
-          const transaction = await connection.getTransaction(
-            signatureInfo.signature,
-            { commitment: "confirmed" }
-          );
-
-          if (transaction) {
-            console.log("New Transaction Details:", transaction);
-
-            for (const instruction of transaction.transaction.message
-              .instructions) {
-              const programIdIndex =
-                instruction.programIdIndex || instruction.programId;
-              const programIdFromInstruction =
-                transaction.transaction.message.accountKeys[programIdIndex];
-
-              if (
-                programIdFromInstruction &&
-                programIdFromInstruction.equals(program.programId)
-              ) {
-                console.log("Instruction for program detected:", instruction);
-
-                let coder = new anchor.BorshInstructionCoder(Gateway_IDL as anchor.Idl);
-                let decodedInstruction = coder.decode(instruction.data, "base58");
-                console.log("Decoded Instruction:", decodedInstruction);
-              }
-            }
-          }
-        } catch (transactionError) {
-          console.error(
-            `Error processing transaction ${signatureInfo.signature}:`,
-            transactionError
-          );
-          // Continue to the next transaction even if an error occurs
-          continue;
-        }
-      }
-    } catch (error) {
-      console.error("Error monitoring new transactions:", error);
-    } finally {
-      // Update lastSignature even if an error occurs
-      if (signatures && signatures.length > 0) {
-        lastSignature = signatures[0].signature;
-      }
-    }
-  }, 1000);
-}
 
 export const initLocalnet = async ({
   port,
@@ -335,7 +257,7 @@ export const initLocalnet = async ({
   port: number;
   exitOnError: boolean;
 }) => {
-  const { gatewayProgram, address } = await setupSolana();
+  const { gatewayProgram, address } = await solanaSetup();
 
   const connection = gatewayProgram.provider.connection;
   await gatewayProgram.methods
@@ -345,7 +267,7 @@ export const initLocalnet = async ({
 
   await new Promise((r) => setTimeout(r, 7000));
 
-  monitorOnlyNewTransactions(gatewayProgram, connection);
+  solanaMonitorTransactions(gatewayProgram, connection);
 
   await new Promise((r) => setTimeout(r, 7000));
